requirements.txt
playwright>=1.47.0
python-dotenv>=1.0.1
pydantic>=2.9.2
rich>=13.9.2
openpyxl>=3.1.5
